Revert "Merge branch 'master' into dyninst_unwind"
This reverts commit 649d26df4fb5dac7435b5f0fc717ebd931e33e49, reversing
changes made to 6fd22a43605c076fc59dc8e2d075defe40ecc435.

+++ src/lib/support/VarMap.hpp
@@ -12,7 +12,7 @@
 // HPCToolkit is at 'hpctoolkit.org' and in 'README.Acknowledgments'.
 // --------------------------------------------------------------------------
 //
-// Copyright ((c)) 2002-2020, Rice University
+// Copyright ((c)) 2002-2019, Rice University
 // All rights reserved.
 //
 // Redistribution and use in source and binary forms, with or without
@@ -57,12 +57,12 @@ private:
   hpcrun_metricVal_t *m_metrics;	   // metric values
   metric_desc_t*      m_list_metric_desc;  // metric descriptions
 
-  int m_num_metrics;
+  size_t m_num_metrics;
   int    m_error_code;
 
 public:
   bool   isVariable(char *expr);
-  double getValue(int var);
+  double getValue(unsigned int var);
   int    getErrorCode();
 
   VarMap(hpcrun_metricVal_t *metrics,
